+++ graphAnalyser/metrics/sizeMetric.cxx
@@ -1,11 +1,10 @@
-#include <string>
 #include "metric.h"
 
 class SizeMetric : public Metric {
     public:
         SizeMetric() :Metric("Size") {
         }
-        int calculate(const Graph& graph) const {
+        double calculate(const Graph& graph) const {
             return graph.getArgs().size();
         };
 };
